feat: transcription with whisper
