+++ whisper_model.py
@@ -1,0 +1,8 @@
+import whisper
+
+language = 'pt'
+model = whisper.load_model('medium')
+
+def transcribe(audio):
+  transcription = model.transcribe(audio, fp16=False, language=language)
+  return transcription['text']
